requirements.txt
python-firebase
requests

